clean up the scripts and reduce error output

+++ scripts_linux/dunst.sh
@@ -1,3 +1,11 @@
-# link or relink dunst stuff
+#!/usr/bin/env bash
+
+if [[ "${DOTFILES}" == "" ]]; then
+    echo "DOTFILES is not set. Exiting to stay safe"
+    exit 1
+fi
+
+test -d ~/.config/dunst && \
 rm -r ~/.config/dunst
+
 ln -s $DOTFILES/config/dunst ~/.config/dunst
