+++ scripts_linux/pcmanfm.sh
@@ -1,3 +1,11 @@
-# link or relink pcmanfm
+#!/usr/bin/env bash
+
+if [[ "${DOTFILES}" == "" ]]; then
+    echo "DOTFILES is not set. Exiting to stay safe"
+    exit 1
+fi
+
+test -d ~/.config/pcmanfm && \
 rm -r ~/.config/pcmanfm
+
 ln -s $DOTFILES/config/pcmanfm ~/.config/pcmanfm
