+++ scripts_linux/pilau.sh
@@ -1,3 +1,9 @@
-# link or relink extras scripts
-rm -r ~/.pilau
+#!/usr/bin/env bash
+
+if [[ "${DOTFILES}" == "" ]]; then
+    echo "DOTFILES is not set. Exiting to stay safe"
+    exit 1
+fi
+
+test -d ~/.pilau && rm -r ~/.pilau
 ln -s $DOTFILES/pilau ~/.pilau
